Examples not clear in 'Adding {wildcard} Requirements'
After passing a test at the Sensio University, I had to verify some routing information.  The result I had with the documentation gave me confusion in the chapter 'Adding {wildcard} Requirements' : 
on lines 58, 160 and 161, it's written that blog_list: ``/blog/{page}`` will match ``/blog/*``;
* blog_show: ``/blog/{slug}`` will *also* match ``/blog/*``. 
This example is not always true, assuming * can mean nothing this will generate a 404 error. Same thing with a slash in the wildcard, if you don't configure the trailing slash it is another error.